Create Unfinished Object Allocator.md
This is some of my new-ish code, unfortunately I have yet to finish it as my time is spent elsewhere lately.

The function of this code is commonly known as 'object allocation' or 'part pooling'. For games that use a lot of the same object it is significantly more performant to initially create a resevoir of objects, pull from that resevoir for the objects when an object is needed to be retrieved, and procedurally create new objects on an as-needed basis.