Create Unfinished Data Replicator.md
This is some of my new-ish code, unfortunately I have yet to finish it as my time is spent elsewhere lately.

The function of this extremely unrefined code is a Data Replicator. Data is hosted on the server and replicated to any clients. 
It's meant to be super lightweight and accessible compared to ReplicaService which is big, comprehensive, and complicated. The intended use-case is mostly just player data replication to other clients.